+++ example/lib/main.dart
@@ -1,6 +1,5 @@
 import 'package:fconnectivity/fconnectivity.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
 
 void main() {
   runApp(const MainApp());
@@ -13,23 +12,39 @@ class MainApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return MaterialApp(
       home: Scaffold(
-        body: BlocProvider<InternetAccessCubit>(
-          create: (context) => InternetAccessCubit(),
-          child: Builder(
-            builder: (context) {
-              return BlocBuilder<InternetAccessCubit, InternetAccessState>(
-                builder: (context, internetAccessState) {
-                  final hasInternetAccess = switch (internetAccessState) {
-                    HasInternetAccess() => true,
-                    HasNoInternetAccess() => false,
-                  };
-                  return Center(
-                    child: Text('Has internet access: $hasInternetAccess'),
-                  );
-                },
-              );
-            },
-          ),
+        body: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            InternetAccessConsumer(
+              onInternetAccessGained: (BuildContext context) => debugPrint(
+                'Internet access gained (logged from InternetAccessConsumer)',
+              ),
+              onInternetAccessLost: (BuildContext context) => debugPrint(
+                'Internet access lost (logged from InternetAccessConsumer)',
+              ),
+              builder: ({
+                required BuildContext context,
+                required bool hasInternetAccess,
+              }) =>
+                  Text(
+                'This text widget rebuilds when internet access state changes. Current state of internet access: ${hasInternetAccess ? 'Available' : 'Not Available'}',
+                textAlign: TextAlign.center,
+              ),
+            ),
+            const SizedBox(height: 20),
+            InternetAccessListener(
+              onInternetAccessGained: (BuildContext context) => debugPrint(
+                'Internet access gained (logged from InternetAccessListener)',
+              ),
+              onInternetAccessLost: (BuildContext context) => debugPrint(
+                'Internet access lost (logged from InternetAccessListener)',
+              ),
+              child: const Text(
+                'This text widget does not rebuild when internet access state changes.',
+                textAlign: TextAlign.center,
+              ),
+            ),
+          ],
         ),
       ),
     );